have markdown for workflows use @example fencing
oops
