+++ docs/src/common_workflows/architecture_search/notebook.jl
@@ -1,6 +1,9 @@
 # # Neural Architecture Search with MLJFlux
 
-# Neural Architecture Search is (NAS) is an instance of hyperparameter tuning concerned
+# This demonstration is available as a Jupyter notebook or julia script
+# [here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/architecture_search).
+
+# Neural Architecture Search (NAS) is an instance of hyperparameter tuning concerned
 # with tuning model hyperparameters defining the architecture itself. Although it's
 # typically performed with sophisticated search algorithms for efficiency, in this example
 # we will be using a simple random search.
@@ -9,9 +12,7 @@ using Pkg     #!md
 Pkg.activate(@__DIR__);     #!md
 Pkg.instantiate();     #!md
 
-# **Julia version** is assumed to be 1.10.* This tutorial is available as a Jupyter
-# notebook or julia script
-# [here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/architecture_search).
+# **Julia version** is assumed to be 1.10.*
 
 # ### Basic Imports
