+++ docs/src/common_workflows/comparison/notebook.md
@@ -4,6 +4,9 @@ EditURL = "notebook.jl"
 
 # Model Comparison with MLJFlux
 
+This demonstration is available as a Jupyter notebook or julia script
+[here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/comparison).
+
 In this workflow example, we see how we can compare different machine learning models
 with a neural network from MLJFlux.
 
@@ -11,7 +14,7 @@ with a neural network from MLJFlux.
 
 ### Basic Imports
 
-````@julia
+````@example comparison
 using MLJ               # Has MLJFlux models
 using Flux              # For more flexibility
 import RDatasets        # Dataset source
@@ -21,7 +24,7 @@ import Optimisers       # native Flux.jl optimisers no longer supported
 
 ### Loading and Splitting the Data
 
-````@julia
+````@example comparison
 iris = RDatasets.dataset("datasets", "iris");
 y, X = unpack(iris, ==(:Species), colname -> true, rng=123);
 nothing #hide
@@ -30,7 +33,7 @@ nothing #hide
 ### Instantiating the models Now let's construct our model. This follows a similar setup
 to the one followed in the [Quick Start](../../index.md#Quick-Start).
 
-````@julia
+````@example comparison
 NeuralNetworkClassifier = @load NeuralNetworkClassifier pkg=MLJFlux
 
 clf1 = NeuralNetworkClassifier(
@@ -44,7 +47,7 @@ clf1 = NeuralNetworkClassifier(
 
 Let's as well load and construct three other classical machine learning models:
 
-````@julia
+````@example comparison
 BayesianLDA = @load BayesianLDA pkg=MultivariateStats
 clf2 = BayesianLDA()
 RandomForestClassifier = @load RandomForestClassifier pkg=DecisionTree
@@ -60,7 +63,7 @@ Instead of just comparing with four models with the default/given hyperparameter
 will give `XGBoostClassifier` an unfair advantage By wrapping it in a `TunedModel` that
 considers the best learning rate η for the model.
 
-````@julia
+````@example comparison
 r1 = range(clf4, :eta, lower=0.01, upper=0.5, scale=:log10)
 tuned_model_xg = TunedModel(
     model=clf4,
@@ -79,7 +82,7 @@ comparing the models over a large set of their hyperparameters.
 
 We simply pass the four models to the `models` argument of the `TunedModel` construct
 
-````@julia
+````@example comparison
 tuned_model = TunedModel(
     models=[clf1, clf2, clf3, tuned_model_xg],
     tuning=Explicit(),
@@ -91,7 +94,7 @@ nothing #hide
 
 Then wrapping our tuned model in a machine and fitting it.
 
-````@julia
+````@example comparison
 mach = machine(tuned_model, X, y);
 fit!(mach, verbosity=0);
 nothing #hide
@@ -99,7 +102,7 @@ nothing #hide
 
 Now let's see the history for more details on the performance for each of the models
 
-````@julia
+````@example comparison
 history = report(mach).history
 history_df = DataFrame(
     mlp = [x[:model] for x in history],
