+++ docs/src/common_workflows/composition/notebook.jl
@@ -1,5 +1,8 @@
 # # Model Composition with MLJFlux
 
+# This demonstration is available as a Jupyter notebook or julia script
+# [here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/composition).
+
 # In this workflow example, we see how MLJFlux enables composing MLJ models with MLJFlux
 # models. We will assume a class imbalance setting and wrap an oversampler with a deep
 # learning model from MLJFlux.
@@ -8,9 +11,7 @@ using Pkg     #!md
 Pkg.activate(@__DIR__);     #!md
 Pkg.instantiate();     #!md
 
-# **Julia version** is assumed to be 1.10.* This tutorial is available as a Jupyter
-# notebook or julia script
-# [here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/composition).
+# **Julia version** is assumed to be 1.10.*
 
 
 # ### Basic Imports
