+++ docs/src/common_workflows/hyperparameter_tuning/notebook.md
@@ -4,13 +4,17 @@ EditURL = "notebook.jl"
 
 # Hyperparameter Tuning with MLJFlux
 
-In this workflow example we learn how to tune different hyperparameters of MLJFlux models with emphasis on training hyperparameters.
+This demonstration is available as a Jupyter notebook or julia script
+[here](https://github.com/FluxML/MLJFlux.jl/tree/dev/docs/src/common_workflows/hyperparameter_tuning).
+
+In this workflow example we learn how to tune different hyperparameters of MLJFlux
+models with emphasis on training hyperparameters.
 
 **Julia version** is assumed to be 1.10.*
 
 ### Basic Imports
 
-````@julia
+````@example hyperparameter_tuning
 using MLJ               # Has MLJFlux models
 using Flux              # For more flexibility
 import RDatasets        # Dataset source
@@ -20,7 +24,7 @@ import Optimisers       # native Flux.jl optimisers no longer supported
 
 ### Loading and Splitting the Data
 
-````@julia
+````@example hyperparameter_tuning
 iris = RDatasets.dataset("datasets", "iris");
 y, X = unpack(iris, ==(:Species), colname -> true, rng=123);
 X = Float32.(X);      # To be compatible with type of network network parameters
@@ -32,7 +36,7 @@ nothing #hide
 Now let's construct our model. This follows a similar setup the one followed in the
 [Quick Start](../../index.md#Quick-Start).
 
-````@julia
+````@example hyperparameter_tuning
 NeuralNetworkClassifier = @load NeuralNetworkClassifier pkg=MLJFlux
 clf = NeuralNetworkClassifier(
     builder=MLJFlux.MLP(; hidden=(5,4), σ=Flux.relu),
@@ -50,7 +54,7 @@ cross-validation.
 
 We start by defining the hyperparameter ranges
 
-````@julia
+````@example hyperparameter_tuning
 r1 = range(clf, :batch_size, lower=1, upper=64)
 etas = [10^x for x in range(-4, stop=0, length=4)]
 optimisers = [Optimisers.Adam(eta) for eta in etas]
@@ -60,7 +64,7 @@ r2 = range(clf, :optimiser, values=optimisers)
 Then passing the ranges along with the model and other arguments to the `TunedModel`
 constructor.
 
-````@julia
+````@example hyperparameter_tuning
 tuned_model = TunedModel(
     model=clf,
     tuning=Grid(goal=25),
@@ -73,7 +77,7 @@ nothing #hide
 
 Then wrapping our tuned model in a machine and fitting it.
 
-````@julia
+````@example hyperparameter_tuning
 mach = machine(tuned_model, X, y);
 fit!(mach, verbosity=0);
 nothing #hide
@@ -81,7 +85,7 @@ nothing #hide
 
 Let's check out the best performing model:
 
-````@julia
+````@example hyperparameter_tuning
 fitted_params(mach).best_model
 ````
 
@@ -92,7 +96,7 @@ hyperparameter of the model
 
 First define the range and wrap it in a learning curve
 
-````@julia
+````@example hyperparameter_tuning
 r = range(clf, :epochs, lower=1, upper=200, scale=:log10)
 curve = learning_curve(
     clf,
@@ -106,7 +110,7 @@ curve = learning_curve(
 
 Then plot the curve
 
-````@julia
+````@example hyperparameter_tuning
 plot(
     curve.parameter_values,
     curve.measurements,
