+++ docs/src/common_workflows/live_training/generate.jl
@@ -1,4 +1,5 @@
-# Execute this julia file to generate the notebooks from ../notebook.jl
+# Execute this julia file in a new julia process to generate the notebooks from
+# ../notebook.jl
 
 joinpath(@__DIR__, "..", "..", "generate.jl") |> include
-generate(@__DIR__, execute=true, pluto=false)
+generate(@__DIR__, execute=false, pluto=false)
