--- docs/src/extended_examples/Spam Detection with RNNs/README.md
@@ -1,15 +1,0 @@
-# Contents
-
-| file                        | description                                              |
-|:----------------------------|:---------------------------------------------------------|
-| `notebook.ipynb`            | Juptyer notebook (executed)                              |
-| `notebook.unexecuted.ipynb` | Jupyter notebook (unexecuted)                            |
-| `notebook.md`               | static markdown (included in MLJFlux.jl docs)            |
-| `notebook.jl`               | executable Julia script annotated with comments          |
-| `generate.jl`               | *maintainers only:* execute to generate first 3 from 4th |
-
-
-# Important 
-
-Scripts or notebooks in this folder cannot be reliably exectued without the accompanying
-Manifest.toml and Project.toml files.
