+++ docs/src/extended_examples/spam_detection/Project.toml
@@ -1,0 +1,13 @@
+[deps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+Languages = "8ef0a80b-9436-5d2c-a485-80b904378c43"
+MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
+MLJFlux = "094fc8d1-fd35-5302-93ea-dabda2abf845"
+MLJText = "5e27fcf9-6bac-46ba-8580-b5712f3d6387"
+Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
+ScientificTypes = "321657f4-b219-11e9-178b-2701a2544e81"
+TextAnalysis = "a2db99b7-8b79-58f8-94bf-bbc811eef33d"
+WordTokenizers = "796a5d58-b03d-544a-977e-18100b691f6e"
